requirements.txt
# requirements.txt
selenium
python-docx
bs4
